+++ driver-core/src/main/java/com/datastax/driver/core/Authenticator.java
@@ -17,18 +17,35 @@ package com.datastax.driver.core;
 
 /**
  * Handles SASL authentication with Cassandra servers.
- *
- * A server which requires authentication responds to a startup
- * message with an challenge in the form of an {@code AuthenticateMessage}.
- * Authenticator implementations should be able to respond to that
- * challenge and perform whatever authentication negotiation is required
- * by the server. The exact nature of that negotiation is specific to the
- * configuration of the server.
+ * <p>
+ * Each time a new connection is created and the server require authentication,
+ * a new instance of this class will be created by the corresponding
+ * {@link AuthProvider} to handler that authentication. The lifecycle of that
+ * new {@code Authenticator} will the be:
+ * <ol>
+ *   <li>The {@code initialResponse} method will be called. The return initial
+ *   value will be send to the server to initiate the handshake.</li>
+ *   <li>The server will respond to each client response by either issuing a
+ *   challenge or indicating that the authentication is complete (successfully or not).
+ *   If a new challenge is issued, the authenticator {@code evaluateChallenge}
+ *   method will be called to produce a response that will be sent to the
+ *   server. This challenge/reponse negotiation will continue until the server
+ *   respond that authentication is succesfull (or an {@code AuthenticationException}
+ *   is raised).
+ *   </li>
+ *   <li>When the server indicates that authentication is successfull, the
+ *   {@code onAuthenticationSucces} method will be call with last informations
+ *   that the server may optionally have sent.
+ *   </li>
+ * </ol>
+ * The exact nature of that negotiation between client and server is specific
+ * to the authentication mechanism configured server side.
  */
 public interface Authenticator {
 
     /**
      * Obtain an initial response token for initializing the SASL handshake
+     *
      * @return the initial response to send to the server, may be null
      */
     public byte[] initialResponse();
@@ -37,9 +54,20 @@ public interface Authenticator {
      * Evaluate a challenge received from the Server. Generally, this method
      * should return null when authentication is complete from the client
      * perspective
+     *
      * @param challenge the server's SASL challenge
      * @return updated SASL token, may be null to indicate the client
      * requires no further action
      */
     public byte[] evaluateChallenge(byte[] challenge);
+
+    /**
+     * Called when authentication is successful with the last information
+     * optionally sent by the server.
+     *
+     * @param token the information sent by the server with the authentication
+     * successful message. This will be {@code null} if the server send no
+     * particular information on authentication success.
+     */
+    public void onAuthenticationSuccess(byte[] token);
 }